refactor(prisma): separate server-side Prisma exports from browser-safe shared types
- Create shared/src/prisma.ts for server-side PrismaClient and model types
- Remove Prisma exports from shared/src/index.ts to prevent Node.js runtime
  from being bundled into frontend builds
- Add shared/prisma subpath export in package.json
- Update prompt-service and worker imports to use shared/prisma
- Consolidate Prisma schema output to shared/src/generated/prisma/
- Remove duplicate worker/prisma/schema.prisma (now uses shared schema)
- Update verify-conventions.sh to exclude generated Prisma files
- Fix backend integration tests: import healthRoutes directly instead of
  full router to prevent /health/health nesting
- Add testing/integration-setup.ts to deep import exclusions

Import pattern:
- Frontend (browser-safe): import { ... } from 'shared'
- Backend (server-side): import { ... } from 'shared/prisma'
