+++ prompt-service/__tests__/unit/services/prompt-resolver.service.test.ts
@@ -19,7 +19,7 @@
 
 import { NotFoundError, ValidationError, ConflictError, isAppError } from '@/errors';
 import { PromptResolverService, promptService, abTestService, type ABTestWithVariants } from '@/services';
-import type { Prompt, PromptVariant, ABVariant, ABVariantConfig } from '@prisma/client';
+import type { Prompt, PromptVariant, ABVariant, ABVariantConfig } from 'shared/prisma';
 
 describe('Error Types', () => {
   describe('NotFoundError', () => {
