+++ prompt-service/prisma/schema.prisma
@@ -1,5 +1,6 @@
 generator client {
-  provider = "prisma-client-js"
+  provider = "prisma-client"
+  output   = "../../shared/src/generated/prisma"  
 }
 
 datasource db {
