+++ prompt-service/src/services/ab-test.service.ts
@@ -13,15 +13,9 @@
  * ESLint disabled rules are due to PrismaLibSql adapter not preserving full PrismaClient types.
  * See: https://github.com/prisma/prisma/issues/21365
  */
-/* eslint-disable @typescript-eslint/no-unsafe-assignment */
-/* eslint-disable @typescript-eslint/no-unsafe-call */
-/* eslint-disable @typescript-eslint/no-unsafe-member-access */
-/* eslint-disable @typescript-eslint/no-unsafe-return */
-/* eslint-disable @typescript-eslint/no-unsafe-argument */
-/* eslint-disable @typescript-eslint/require-await */
 
 import {prisma} from '@/prisma';
-import type {ABTest, ABVariant, ABVariantConfig, Prisma} from '@prisma/client';
+import type { Prisma, ABTest, ABVariant, ABVariantConfig } from 'shared/prisma';
 import {NotFoundError, ValidationError} from '@/errors';
 
 export interface ABTestWithVariants extends ABTest {
