+++ worker/package.json
@@ -9,7 +9,6 @@
     "build": "tsup",
     "clean": "rm -rf dist tsconfig.tsbuildinfo",
     "start": "node dist/index.js",
-    "db:generate": "prisma generate",
     "lint": "eslint src/",
     "lint:fix": "eslint src/ --fix",
     "type-check": "tsc --noEmit",
